refactor(web): introduce the mutation-hook layer (P4, orchestrators)
Establish web/src/hooks/mutations/ as the write-side boundary, following
the TanStack "Mastering Mutations" split: each per-operation hook owns only
the data-consistency logic that must always run (cache invalidation, plus
best-effort domain follow-ups), while toast/navigate/component-state resets
stay on the .mutate() call at the component so they skip when unmounted.

- Move the 4 existing triad-owner hooks (useAddRepoCollaborator,
  useRemoveRepoCollaborator, useReuseAssignment, useAcceptAndVerifyMembership)
  into hooks/mutations/.
- Extract the multi-call orchestrators into named hooks: useAddStaffMember,
  useEnrollOrInviteStudent, useClaimInstructor, useSaveServiceToken,
  useSyncRoster, useMigrateRoster, useReinviteFailedInvite, usePlanTeardown,
  useExecuteTeardown. Pages shed ~500 lines of inline write-wiring.

Deliberate variants preserved, not forced into a uniform mold: the Teardown
ConfirmModal re-throw contract (execute still rejects), migrate's no-invalidate
mount-fired path, AddStudent's optimistic seed-and-reconcile, and useSafeSubmit
composition. Trivial single-call writes and parent-delegated/step-streaming
sites left as-is per the survey + TanStack over-abstraction guidance.

No behavior change; the github-core no-cycle guard stays clean (hooks import
downward only).
