+++ web/src/components/modals/GroupCollaboratorsModal.tsx
@@ -9,8 +9,8 @@ import { Alert, AnimatedAlert, Button, Modal } from "@/components/ui"
 import { useGithubAuth } from "@/auth/useGithubAuth"
 import useGetRepo from "@/hooks/useGetRepo"
 import useGetRepoCollaborators from "@/hooks/useGetRepoCollaborators"
-import useAddRepoCollaborator from "@/hooks/useAddRepoCollaborator"
-import useRemoveRepoCollaborator from "@/hooks/useRemoveRepoCollaborator"
+import useAddRepoCollaborator from "@/hooks/mutations/useAddRepoCollaborator"
+import useRemoveRepoCollaborator from "@/hooks/mutations/useRemoveRepoCollaborator"
 import { getName } from "@/util/students"
 import { GitHubAPIError } from "@/github-core/errors"
 import type { Student } from "@/types/classroom"
