+++ web/src/pages/OnboardingPage.tsx
@@ -18,7 +18,7 @@ import { useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useGithubAuth } from "@/auth/useGithubAuth"
 import useGetOwnOrgMembership from "@/hooks/useGetOwnOrgMembership"
-import { useAcceptAndVerifyMembership } from "@/hooks/useAcceptAndVerifyMembership"
+import { useAcceptAndVerifyMembership } from "@/hooks/mutations/useAcceptAndVerifyMembership"
 import { isMembershipReadError } from "@/util/membershipReadError"
 import { deriveOnboardingState } from "@/util/onboardingState"
 import { EnterDiv } from "@/lib/motionComponents"
